requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
flask-migrate
gunicorn
redis
rq
# psycopg2
requests
jinja2
